feat: Integrate schematic viewer
