+++ package.json
@@ -61,6 +61,7 @@
     "@tscircuit/mm": "^0.0.8",
     "@tscircuit/pcb-viewer": "^1.10.5",
     "@tscircuit/props": "^0.0.77",
+    "@tscircuit/schematic-viewer": "^1.4.2",
     "@types/file-saver": "^2.0.7",
     "@types/ms": "^0.7.34",
     "@typescript/ata": "^0.9.7",
@@ -108,7 +109,7 @@
     "@babel/standalone": "^7.25.6",
     "@biomejs/biome": "^1.9.2",
     "@playwright/test": "^1.48.0",
-    "@tscircuit/core": "^0.0.132",
+    "@tscircuit/core": "^0.0.133",
     "@tscircuit/prompt-benchmarks": "^0.0.14",
     "@types/babel__standalone": "^7.1.7",
     "@types/bun": "^1.1.10",
