+++ src/lib/templates/blank-circuit-board-template.ts
@@ -7,12 +7,14 @@ export default () => (
       resistance="1k"
       footprint="0402"
       name="R1"
+      schX={3}
       pcbX={3}
     />
     <capacitor
       capacitance="1000pF"
       footprint="0402"
       name="C1"
+      schX={-3}
       pcbX={-3}
     />
     <trace from=".R1 > .pin1" to=".C1 > .pin1" />
