+++ src/lib/templates/blank-footprint-template.ts
@@ -7,6 +7,8 @@ export const MyFootprint = (props: any) => (
       shape="rect"
       width="0.5mm"
       height="0.5mm"
+      schX={-1}
+      schY={0}
       pcbX={-1}
       pcbY={0}
       portHints={["1"]}
@@ -15,6 +17,8 @@ export const MyFootprint = (props: any) => (
       shape="rect"
       width="0.5mm"
       height="0.5mm"
+      schX={1}
+      schY={0}
       pcbX={1}
       pcbY={0}
       portHints={["2"]}
